+++ bin/config/global_config.py
@@ -53,6 +53,7 @@ transformation_choices = ['MCAS', 'JMESPATH']
     # LogLevels
 log_levels_available = ['DEBUG', 'INFO', 'WARNING', 'ERROR', 'CRITICAL']
 log_level_default = 'WARNING'
+log_debugloglines_default = False
 
 # INPUT Configuration
 input_rerun_retries = 3                         # Number of rerun attempts before giving up